requirements.txt
requests
python-dotenv
openai
llama-index
pandas
